+++ sql/03_load_data.sql
@@ -9,6 +9,7 @@ SET search_path TO gastronomia_caba;
 \copy dim_organizador FROM 'data/processed/dim_organizador.csv' WITH (FORMAT csv, HEADER true, ENCODING 'UTF8');
 
 \copy fact_establecimiento FROM 'data/processed/fact_establecimiento.csv' WITH (FORMAT csv, HEADER true, ENCODING 'UTF8');
+\copy fact_habilitacion_gastronomica FROM 'data/processed/fact_habilitacion_gastronomica.csv' WITH (FORMAT csv, HEADER true, ENCODING 'UTF8');
 \copy fact_evento_gastronomico FROM 'data/processed/fact_evento_gastronomico.csv' WITH (FORMAT csv, HEADER true, ENCODING 'UTF8');
 \copy fact_programa_politica FROM 'data/processed/fact_programa_politica.csv' WITH (FORMAT csv, HEADER true, ENCODING 'UTF8');
 \copy fact_mercado_feria FROM 'data/processed/fact_mercado_feria.csv' WITH (FORMAT csv, HEADER true, ENCODING 'UTF8');